+++ crates/rattler/src/install/link_script.rs
@@ -34,19 +34,19 @@ impl LinkScriptType {
         if platform.is_windows() {
             match self {
                 LinkScriptType::PreUnlink => {
-                    format!("Scripts/.{}-pre-unlink.bat", name)
+                    format!("Scripts/.{name}-pre-unlink.bat")
                 }
                 LinkScriptType::PostLink => {
-                    format!("Scripts/.{}-post-link.bat", name)
+                    format!("Scripts/.{name}-post-link.bat")
                 }
             }
         } else {
             match self {
                 LinkScriptType::PreUnlink => {
-                    format!("bin/.{}-pre-unlink.sh", name)
+                    format!("bin/.{name}-pre-unlink.sh")
                 }
                 LinkScriptType::PostLink => {
-                    format!("bin/.{}-post-link.sh", name)
+                    format!("bin/.{name}-post-link.sh")
                 }
             }
         }
@@ -90,10 +90,16 @@ pub fn run_link_scripts<'a>(
             env.insert("PKG_BUILDNUM".to_string(), prec.build_number.to_string());
 
             let shell = if platform.is_windows() {
-                ShellEnum::CmdExe(CmdExe::default())
+                ShellEnum::CmdExe(CmdExe)
             } else {
-                ShellEnum::Bash(Bash::default())
+                ShellEnum::Bash(Bash)
             };
+
+            tracing::info!(
+                "Running {} script for {}",
+                link_script_type.to_string(),
+                prec.name.as_normalized()
+            );
 
             match rattler_shell::run_in_environment(
                 target_prefix,
@@ -116,9 +122,9 @@ pub fn run_link_scripts<'a>(
                     prec.name.as_normalized(),
                     message
                 );
+                // Remove the message file
+                std::fs::remove_file(&message_file)?;
             }
-            // Remove the message file
-            std::fs::remove_file(&message_file)?;
         }
     }
 
@@ -163,5 +169,78 @@ impl InstallDriver {
             target_prefix,
             &transaction.platform,
         )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use rattler_conda_types::{Platform, PrefixRecord, RepoDataRecord};
+
+    use crate::{
+        get_repodata_record,
+        install::{
+            test_utils::execute_transaction, transaction, InstallDriver, InstallOptions,
+            TransactionOperation,
+        },
+        package_cache::PackageCache,
+    };
+
+    fn test_operations() -> Vec<TransactionOperation<PrefixRecord, RepoDataRecord>> {
+        let repodata_record_1 =
+            get_repodata_record("link-scripts/link-scripts-0.1.0-h4616a5c_0.tar.bz2");
+
+        vec![TransactionOperation::Install(repodata_record_1)]
+    }
+
+    #[tokio::test]
+    async fn test_run_link_scripts() {
+        let target_prefix = tempfile::tempdir().unwrap();
+
+        let operations = test_operations();
+
+        let transaction = transaction::Transaction::<PrefixRecord, RepoDataRecord> {
+            operations,
+            python_info: None,
+            current_python_info: None,
+            platform: Platform::current(),
+        };
+
+        let packages_dir = tempfile::tempdir().unwrap();
+        let cache = PackageCache::new(packages_dir.path());
+
+        execute_transaction(
+            transaction,
+            target_prefix.path(),
+            &reqwest_middleware::ClientWithMiddleware::from(reqwest::Client::new()),
+            &cache,
+            &InstallDriver::default(),
+            &InstallOptions::default(),
+        )
+        .await;
+
+        // check that the post-link script was run
+        assert!(target_prefix.path().join("i-was-post-linked").exists());
+
+        // unlink the package
+        let prefix_records = PrefixRecord::collect_from_prefix(target_prefix.path()).unwrap();
+        let transaction = transaction::Transaction::<PrefixRecord, RepoDataRecord> {
+            operations: vec![TransactionOperation::Remove(prefix_records[0].clone())],
+            python_info: None,
+            current_python_info: None,
+            platform: Platform::current(),
+        };
+
+        execute_transaction(
+            transaction,
+            target_prefix.path(),
+            &reqwest_middleware::ClientWithMiddleware::from(reqwest::Client::new()),
+            &cache,
+            &InstallDriver::default(),
+            &InstallOptions::default(),
+        )
+        .await;
+
+        // check that the pre-unlink script was run
+        assert!(!target_prefix.path().join("i-was-post-linked").exists());
     }
 }
