+++ crates/rattler/src/install/mod.rs
@@ -23,6 +23,9 @@ mod python;
 mod transaction;
 pub mod unlink;
 
+#[cfg(test)]
+mod test_utils;
+
 pub use crate::install::entry_point::{get_windows_launcher, python_entry_point_template};
 pub use driver::InstallDriver;
 pub use link::{link_file, LinkFileError};
