+++ test-data/link-scripts/recipe/post-link.bat
@@ -1,0 +1,3 @@
+echo. > %PREFIX%\i-was-post-linked
+
+echo This is a post-link script that just ran for the recipe package. >> %PREFIX%\.messages.txt
