+++ test-data/link-scripts/recipe/post-link.sh
@@ -1,0 +1,5 @@
+touch $PREFIX/i-was-post-linked
+
+cat >> ${PREFIX}/.messages.txt << EOF
+This is a post-link script that just ran for the recipe package.
+EOF
