+++ test-data/link-scripts/recipe/pre-unlink.bat
@@ -1,0 +1,1 @@
+rm %PREFIX%/i-was-post-linked
