+++ test-data/link-scripts/recipe/pre-unlink.sh
@@ -1,0 +1,1 @@
+rm $PREFIX/i-was-post-linked
